Reestruturando dinâmicas de View e Routes

+++ model/fabrics.go
@@ -4,10 +4,10 @@ type Tecido struct {
 	ID uint
 	Nome string
 	FichaTecnica string
-	Rendimento float64 
+	Rendimento int 
 	Aproveitamento string
 	Largura string
-	Custo float64
+	Custo int 
 	Observacoes string
 }
 
@@ -15,7 +15,7 @@ type Aviamento struct {
 	ID            int     `json:"id"`
 	Tipo          string  `json:"tipo"`
 	Cor           string  `json:"cor"`
-	CustoUnitario float64 `json:"custounitario"`
+	CustoUnitario int`json:"custounitario"`
 	QtdAviamento  uint    `json:"qtdaviamento"`
 }
 
@@ -23,6 +23,6 @@ type Desenho struct {
 	Tipo          string  `json:"tipo"`
 	Imagem        string  `json:"imagem"`
 	Posicao       string  `json:"posicao"`
-	CustoUnitario float64 `json:"custounitario"`
+	CustoUnitario int`json:"custounitario"`
 }
