+++ model/ficha_produto.go
@@ -3,12 +3,12 @@ package model
 // FichaProduto representa a estrutura da ficha técnica do Produto final. Ela herda dados da FichaModelo para a criação de cada produto de forma personalizada 
 type FichaProduto struct {
 	ID              int         `json:"id"`
-	FichaModeloID     int       `json:"fichamodeloid"`
+	FichaModeloID   int       `json:"fichamodeloid"`
 	Cliente         string      `json:"cliente"`
 	Tecidos         []Tecido    `json:"tecidos"`
-	CustoTecido     float64     `json:"custotecido"`
+	CustoTecido     int 	    `json:"custotecido"`
 	Aviamentos      []Aviamento `json:"aviamentos"`
-	CustoAviamentos float64     `json:"custoaviamentos"`
+	CustoAviamentos int 	    `json:"custoaviamentos"`
 	Desenhos        []Desenho   `json:"desenhos"`
-	CustoDesenho    float64     `json:"custodesenho"`
+	CustoDesenho    int     `json:"custodesenho"`
 }
